+++ app/src/main/AndroidManifest.xml
@@ -7,14 +7,14 @@
     <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE"/>
 
     <application
-        android:name=".login.TrainingApplication"
+        android:name=".training.TrainingApplication"
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
         android:theme="@style/AppTheme">
 
         <activity
-            android:name=".login.ui.training.LoginActivity"
+            android:name=".training.ui.login.LoginActivity"
             android:screenOrientation="portrait">
 
             <intent-filter>
@@ -25,7 +25,7 @@
         </activity>
 
         <activity
-            android:name=".login.ui.viewpager.ViewpagerActivity"
+            android:name=".training.ui.viewpager.ViewpagerActivity"
             android:screenOrientation="portrait" />
 
     </application>
