--- app/src/main/java/ar/com/wolox/android/login/BaseConfiguration.java
@@ -1,8 +1,0 @@
-package ar.com.wolox.android.login;
-
-class BaseConfiguration {
-
-    public static final String EXAMPLE_CONFIGURAITON_KEY = "https://jsonplaceholder.typicode.com";
-    public static final String SHARED_PREFERENCES_NAME = "private-shared-prefs";
-
-}
