--- app/src/main/java/ar/com/wolox/android/login/model/ExampleModel.java
@@ -1,8 +1,0 @@
-package ar.com.wolox.android.login.model;
-
-
-public class ExampleModel {
-
-    public int someNumber;
-
-}
