--- app/src/main/java/ar/com/wolox/android/login/model/Post.java
@@ -1,15 +1,0 @@
-package ar.com.wolox.android.login.model;
-
-public class Post {
-
-    private String title;
-    private String body;
-
-    public String getTitle() {
-        return title;
-    }
-
-    public String getBody() {
-        return body;
-    }
-}
