--- app/src/main/java/ar/com/wolox/android/login/model/ViewPagerModel.java
@@ -1,7 +1,0 @@
-package ar.com.wolox.android.login.model;
-
-public class ViewPagerModel {
-
-    public String Name;
-
-}
