--- app/src/main/java/ar/com/wolox/android/login/ui/training/ILoginView.java
@@ -1,8 +1,0 @@
-package ar.com.wolox.android.login.ui.training;
-
-public interface ILoginView {
-
-    void onLoginSuccess();
-    void onLoginEmailError(String error);
-    void onLoginPasswordError(String error);
-}
