--- app/src/main/java/ar/com/wolox/android/login/ui/training/LoginActivity.java
@@ -1,23 +1,0 @@
-package ar.com.wolox.android.login.ui.training;
-
-
-import ar.com.wolox.android.R;
-import ar.com.wolox.wolmo.core.activity.WolmoActivity;
-
-import javax.inject.Inject;
-
-public class LoginActivity extends WolmoActivity {
-
-    @Inject
-    LoginFragment mLoginFragment;
-
-    @Override
-    protected int layout() {
-        return R.layout.activity_base;
-    }
-
-    @Override
-    protected void init() {
-        replaceFragment(R.id.activity_base_content, mLoginFragment);
-    }
-}
