--- app/src/main/java/ar/com/wolox/android/login/ui/viewpager/ViewPagerActivitySubcomponent.java
@@ -1,14 +1,0 @@
-package ar.com.wolox.android.login.ui.viewpager;
-
-import ar.com.wolox.android.login.ui.viewpager.fragment.ViewPagerFragmentModule;
-
-import dagger.Subcomponent;
-import dagger.android.AndroidInjector;
-
-@Subcomponent(modules = {ViewPagerFragmentModule.class})
-public interface ViewPagerActivitySubcomponent extends AndroidInjector<ViewpagerActivity> {
-
-    @Subcomponent.Builder
-    abstract class Builder extends AndroidInjector.Builder<ViewpagerActivity> {
-    }
-}
