--- app/src/main/java/ar/com/wolox/android/login/ui/viewpager/fragment/ViewPagerFragmentSubComponent.java
@@ -1,12 +1,0 @@
-package ar.com.wolox.android.login.ui.viewpager.fragment;
-
-import dagger.Subcomponent;
-import dagger.android.AndroidInjector;
-
-@Subcomponent
-public interface ViewPagerFragmentSubComponent extends AndroidInjector<ViewPagerFragment> {
-
-    @Subcomponent.Builder
-    abstract class Builder extends AndroidInjector.Builder<ViewPagerFragment> {
-    }
-}
