--- app/src/main/java/ar/com/wolox/android/login/ui/viewpager/random/IRandomView.java
@@ -1,8 +1,0 @@
-package ar.com.wolox.android.login.ui.viewpager.random;
-
-public interface IRandomView {
-
-    void setUsername(String username);
-
-    void onRandomNumberUpdate(int someNumber);
-}
