--- app/src/main/java/ar/com/wolox/android/login/ui/viewpager/request/IRequestView.java
@@ -1,10 +1,0 @@
-package ar.com.wolox.android.login.ui.viewpager.request;
-
-public interface IRequestView {
-
-    void setNewsTitle(String title);
-
-    void setNewsBody(String body);
-
-    void showError();
-}
