--- app/src/main/java/ar/com/wolox/android/login/utils/Extras.java
@@ -1,19 +1,0 @@
-package ar.com.wolox.android.login.utils;
-
-import android.app.Activity;
-import android.content.SharedPreferences;
-import android.support.v4.app.Fragment;
-
-/**
- * Util class to store keys to use with {@link SharedPreferences} or as argument between
- * {@link Fragment} or {@link Activity}.
- */
-public class Extras {
-
-    private Extras() {}
-
-    public static class UserLogin {
-        public static final String EMAIL = "email";
-        public static final String PASSWORD = "password";
-    }
-}
