+++ app/src/main/java/ar/com/wolox/android/training/BaseConfiguration.java
@@ -1,0 +1,8 @@
+package ar.com.wolox.android.training;
+
+class BaseConfiguration {
+
+    public static final String EXAMPLE_CONFIGURAITON_KEY = "https://jsonplaceholder.typicode.com";
+    public static final String SHARED_PREFERENCES_NAME = "private-shared-prefs";
+
+}
