+++ app/src/main/java/ar/com/wolox/android/training/model/ExampleModel.java
@@ -1,0 +1,8 @@
+package ar.com.wolox.android.training.model;
+
+
+public class ExampleModel {
+
+    public int someNumber;
+
+}
