+++ app/src/main/java/ar/com/wolox/android/training/model/Post.java
@@ -1,0 +1,15 @@
+package ar.com.wolox.android.training.model;
+
+public class Post {
+
+    private String title;
+    private String body;
+
+    public String getTitle() {
+        return title;
+    }
+
+    public String getBody() {
+        return body;
+    }
+}
