+++ app/src/main/java/ar/com/wolox/android/training/model/ViewPagerModel.java
@@ -1,0 +1,7 @@
+package ar.com.wolox.android.training.model;
+
+public class ViewPagerModel {
+
+    public String Name;
+
+}
