+++ app/src/main/java/ar/com/wolox/android/training/ui/login/ILoginView.java
@@ -1,0 +1,8 @@
+package ar.com.wolox.android.training.ui.login;
+
+public interface ILoginView {
+
+    void onLoginSuccess();
+    void onLoginEmailError(String error);
+    void onLoginPasswordError(String error);
+}
