+++ app/src/main/java/ar/com/wolox/android/training/ui/viewpager/fragment/ViewPagerFragmentSubComponent.java
@@ -1,0 +1,12 @@
+package ar.com.wolox.android.training.ui.viewpager.fragment;
+
+import dagger.Subcomponent;
+import dagger.android.AndroidInjector;
+
+@Subcomponent
+public interface ViewPagerFragmentSubComponent extends AndroidInjector<ViewPagerFragment> {
+
+    @Subcomponent.Builder
+    abstract class Builder extends AndroidInjector.Builder<ViewPagerFragment> {
+    }
+}
