+++ app/src/main/java/ar/com/wolox/android/training/ui/viewpager/random/IRandomView.java
@@ -1,0 +1,8 @@
+package ar.com.wolox.android.training.ui.viewpager.random;
+
+public interface IRandomView {
+
+    void setUsername(String username);
+
+    void onRandomNumberUpdate(int someNumber);
+}
