+++ app/src/main/java/ar/com/wolox/android/training/ui/viewpager/request/IRequestView.java
@@ -1,0 +1,10 @@
+package ar.com.wolox.android.training.ui.viewpager.request;
+
+public interface IRequestView {
+
+    void setNewsTitle(String title);
+
+    void setNewsBody(String body);
+
+    void showError();
+}
