--- app/src/production/java/ar/com/wolox/android/login/Configuration.java
@@ -1,5 +1,0 @@
-package ar.com.wolox.android.login;
-
-public class Configuration extends BaseConfiguration {
-    public static final String API_ENDPOINT = "http://api.wolox.com.ar";
-}
