+++ app/src/production/java/ar/com/wolox/android/training/Configuration.java
@@ -1,5 +1,5 @@
 package ar.com.wolox.android.training;
 
 public class Configuration extends BaseConfiguration {
-    public static final String API_ENDPOINT = "http://stage.api.wolox.com.ar";
+    public static final String API_ENDPOINT = "http://api.wolox.com.ar";
 }
